feat: Add TodoDetail screen component
This commit adds the TodoDetail screen component, which displays the details of a specific todo item. The component imports the necessary dependencies from react-native and renders a simple view with a text component displaying "TodoDetail". This screen will be used for navigating to the details of a todo item within the app.