feat: Add TodoItem component
This commit adds the TodoItem component, which is responsible for rendering a single todo item in the app. The component imports the necessary dependencies from react-native and renders a view with a text component displaying "TodoItem". This component will be used within the Home screen to display the list of todo items.